Fix all clippy warnings across the codebase
- Replace `loop { if let ... else { break } }` with `while let` (balance.rs, kwayfm.rs)
- Use `.clamp()` instead of `.max().min()` chains (balance.rs, fm.rs)
- Change `&mut Vec<T>` to `&mut [T]` in function signatures (kwayrefine.rs, refine2way.rs)
- Add `#[allow(clippy::too_many_arguments)]` for C-ported functions (kwayfm.rs)
- Convert `for i in 0..n` to slice iteration where loop var only indexes one array
- Add `#[allow(clippy::needless_range_loop)]` where index is used across multiple arrays
- Remove dead `newcut` assignment in balance rollback (balance.rs)
- Suppress dead_code warnings for `call_count` and `contains` utility methods
- Fix doc formatting issues (fm.rs, pmetis.rs)
- Apply cargo fmt

+++ src/graph/contract.rs
@@ -30,7 +30,7 @@ pub fn create_coarse_graph(
     let mut cedge_weights = vec![0 as Idx; max_edges];
 
     // Hash table for edge merging
-    let mut htable = vec![-1 as i32; mask + 1];
+    let mut htable = vec![-1_i32; mask + 1];
 
     let mut cnum_edges: usize = 0;
